requirements.txt
Flask
flask-cors
psycopg2-binary
gunicorn
python-dotenv
Flask-JWT-Extended
Werkzeug